+++ main_io.h
@@ -1,10 +1,10 @@
 #include "mmq.h"
 
 //funcao que le n e k
-void read_n_k(int *n, int *k);
+void read_n_k(long long int *n, long long int *k);
 
 //funcao que le os pontos
-points_t *read_points(int n);
+points_t *read_points(long long int n);
 
 //funcao que libera memoria dos pontos
 void free_points(points_t *points);
