+++ utils.h
@@ -30,6 +30,6 @@ typedef double rtime_t;
 
 /* Funções */
 rtime_t timestamp(void);
-string_t markerName(string_t baseName, int n);
+string_t markerName(string_t baseName, long long int n);
 
 #endif /* __UTILS_H__ */
